Feat: Contract utils functions + Diamond update

+++ packages/libs/contracts-sdk/contracts/libs/LibVincentAppFacet.sol
@@ -11,7 +11,7 @@ library LibVincentAppFacet {
      * @param appId Unique identifier for the newly registered app
      * @param manager Address of the app manager
      */
-    event NewAppRegistered(uint256 indexed appId, address indexed manager);
+    event NewAppRegistered(uint40 indexed appId, address indexed manager);
 
     /**
      * @notice Emitted when a new app version is registered
@@ -19,7 +19,7 @@ library LibVincentAppFacet {
      * @param appVersion Version number of the newly registered app version
      * @param manager Address of the app manager who registered the new version
      */
-    event NewAppVersionRegistered(uint256 indexed appId, uint256 indexed appVersion, address indexed manager);
+    event NewAppVersionRegistered(uint40 indexed appId, uint24 indexed appVersion, address indexed manager);
 
     /**
      * @notice Emitted when an app version's enabled status is changed
@@ -27,21 +27,21 @@ library LibVincentAppFacet {
      * @param appVersion Version number of the app being enabled/disabled
      * @param enabled New enabled status of the app version
      */
-    event AppEnabled(uint256 indexed appId, uint256 indexed appVersion, bool indexed enabled);
+    event AppEnabled(uint40 indexed appId, uint24 indexed appVersion, bool indexed enabled);
 
     /**
      * @notice Emitted when a new delegatee is added to an app
      * @param appId ID of the app
      * @param delegatee Address of the delegatee added to the app
      */
-    event DelegateeAdded(uint256 indexed appId, address indexed delegatee);
+    event DelegateeAdded(uint40 indexed appId, address indexed delegatee);
 
     /**
      * @notice Emitted when a delegatee is removed from an app
      * @param appId ID of the app
      * @param delegatee Address of the delegatee removed from the app
      */
-    event DelegateeRemoved(uint256 indexed appId, address indexed delegatee);
+    event DelegateeRemoved(uint40 indexed appId, address indexed delegatee);
 
     /**
      * @notice Emitted when a new lit action is registered
@@ -53,20 +53,20 @@ library LibVincentAppFacet {
      * @notice Emitted when an app is deleted
      * @param appId ID of the deleted app
      */
-    event AppDeleted(uint256 indexed appId);
+    event AppDeleted(uint40 indexed appId);
 
     /**
      * @notice Emitted when an app is undeleted
      * @param appId ID of the undeleted app
      */
-    event AppUndeleted(uint256 indexed appId);
+    event AppUndeleted(uint40 indexed appId);
 
     /**
      * @notice Error thrown when a non-manager attempts to modify an app
      * @param appId ID of the app being modified
      * @param msgSender Address that attempted the unauthorized modification
      */
-    error NotAppManager(uint256 appId, address msgSender);
+    error NotAppManager(uint40 appId, address msgSender);
 
     /**
      * @notice Error thrown when ability and policy array lengths don't match
@@ -80,14 +80,14 @@ library LibVincentAppFacet {
      * @param appId ID of the app the delegatee is already registered to
      * @param delegatee Address of the delegatee that is already registered
      */
-    error DelegateeAlreadyRegisteredToApp(uint256 appId, address delegatee);
+    error DelegateeAlreadyRegisteredToApp(uint40 appId, address delegatee);
 
     /**
      * @notice Error thrown when trying to remove a delegatee not registered to the specified app
      * @param appId ID of the app from which removal was attempted
      * @param delegatee Address of the delegatee that is not registered to the app
      */
-    error DelegateeNotRegisteredToApp(uint256 appId, address delegatee);
+    error DelegateeNotRegisteredToApp(uint40 appId, address delegatee);
 
     /**
      * @notice Error thrown when trying to set app version enabled status to its current status
@@ -95,28 +95,28 @@ library LibVincentAppFacet {
      * @param appVersion Version number of the app
      * @param enabled Current enabled status
      */
-    error AppVersionAlreadyInRequestedState(uint256 appId, uint256 appVersion, bool enabled);
+    error AppVersionAlreadyInRequestedState(uint40 appId, uint24 appVersion, bool enabled);
 
     /**
      * @notice Error thrown when trying to use an empty policy IPFS CID
      * @param appId ID of the app
      * @param abilityIndex Index of the ability in the abilities array
      */
-    error EmptyPolicyIpfsCidNotAllowed(uint256 appId, uint256 abilityIndex);
+    error EmptyPolicyIpfsCidNotAllowed(uint40 appId, uint256 abilityIndex);
 
     /**
      * @notice Error thrown when an ability IPFS CID is empty
      * @param appId ID of the app
      * @param abilityIndex Index of the ability in the abilities array
      */
-    error EmptyAbilityIpfsCidNotAllowed(uint256 appId, uint256 abilityIndex);
+    error EmptyAbilityIpfsCidNotAllowed(uint40 appId, uint256 abilityIndex);
 
     /**
      * @notice Error thrown when an ability IPFS CID is present more than once
      * @param appId ID of the app
      * @param abilityIndex Index of the ability in the abilities array
      */
-    error DuplicateAbilityIpfsCidNotAllowed(uint256 appId, uint256 abilityIndex);
+    error DuplicateAbilityIpfsCidNotAllowed(uint40 appId, uint256 abilityIndex);
 
     /**
      * @notice Error thrown when a policy IPFS CID is present more than once
@@ -124,7 +124,7 @@ library LibVincentAppFacet {
      * @param abilityIndex Index of the ability in the abilities array
      * @param policyIndex Index of the policy in the policies array
      */
-    error DuplicateAbilityPolicyIpfsCidNotAllowed(uint256 appId, uint256 abilityIndex, uint256 policyIndex);
+    error DuplicateAbilityPolicyIpfsCidNotAllowed(uint40 appId, uint256 abilityIndex, uint256 policyIndex);
 
     /**
      * @notice Error thrown when a delegatee address is the zero address
@@ -135,14 +135,14 @@ library LibVincentAppFacet {
      * @notice Error thrown when no abilities are provided during app version registration
      * @param appId ID of the app
      */
-    error NoAbilitiesProvided(uint256 appId);
+    error NoAbilitiesProvided(uint40 appId);
 
     /**
      * @notice Error thrown when no policies are provided for an ability
      * @param appId ID of the app
      * @param abilityIndex Index of the ability with no policies
      */
-    error NoPoliciesProvidedForAbility(uint256 appId, uint256 abilityIndex);
+    error NoPoliciesProvidedForAbility(uint40 appId, uint256 abilityIndex);
 
     /**
      * @notice Error thrown when the top-level ability arrays have mismatched lengths
@@ -177,20 +177,20 @@ library LibVincentAppFacet {
      * @notice Error thrown when the app is already deleted
      * @param appId ID of the deleted app
      */
-    error AppAlreadyDeleted(uint256 appId);
+    error AppAlreadyDeleted(uint40 appId);
 
     /**
      * @notice Error thrown when the app is already undeleted
      * @param appId ID of the undeleted app
      */
-    error AppAlreadyUndeleted(uint256 appId);
+    error AppAlreadyUndeleted(uint40 appId);
 
     /**
      * @notice Error thrown when the app version has delegated agents
      * @param appId ID of the app
      * @param appVersion Version number of the app
      */
-    error AppVersionHasDelegatedAgents(uint256 appId, uint256 appVersion);
+    error AppVersionHasDelegatedAgents(uint40 appId, uint24 appVersion);
 
     /**
      * @notice Error thrown when the app ID is zero
@@ -201,5 +201,5 @@ library LibVincentAppFacet {
      * @notice Error thrown when the app is already registered
      * @param appId ID of the app
      */
-    error AppAlreadyRegistered(uint256 appId);
+    error AppAlreadyRegistered(uint40 appId);
 }
